StatScraper
Beefy py script that scrapes (gouges) stats from stats.nba.com. The most useful features for our purposes would be 'get_games_today' and 'get_game_stats,' which return game ids for the games played today and a dict of all the final stats for every player on each team, respectively (you can see the results in the 'games' collection on Atlas). I included a brief Readme at the top of the file in case you're inclined to try it out.